config.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# --- API & LLM CONFIGURATION ---
API_KEY = os.getenv("OPENROUTER_API_KEY")
BASE_URL = "https://openrouter.ai/api/v1"
# Arcee AI model provides <3s latency with large context windows
LLM_MODEL = "arcee-ai/trinity-large-preview:free"

# --- RAG CONFIGURATION ---
DB_FOLDER = "./chroma_db"
COLLECTION_NAME = "university_data"
EMBEDDING_MODEL = "BAAI/bge-small-en-v1.5"
DATA_FOLDER = "Data"
TOP_K_DOCS = 10
MASTER_REPORT_PATH = os.path.join(DATA_FOLDER, "CUTMAP_FINAL_COMPLETE_REPORT.md")

# Intent routing keywords
BROAD_QUERY_KEYWORDS = [
    "all programs", "what programs", "programs offered", "courses offered",
    "all courses", "list of programs", "list of courses", "what courses",
    "which programs", "which courses", "all departments", "all schools",
    "overview", "all facilities", "list of faculty", "all faculty",
    "all staff", "total programs", "how many programs", "what do you offer",
    "what can i study", "available programs", "available courses",
    "all branches", "all specializations", "types of programs",
]

# --- SYSTEM PROMPTS ---
SYSTEM_INSTRUCTION = """
You are **Centia**, the incredibly warm, friendly, and deeply knowledgeable voice assistant for Centurion University (CUTM-AP).

### YOUR PERSONALITY & TONE:
- Speak exactly like a helpful, extroverted human university guide.
- Use natural conversational transitions like "Well", "Actually", or "By the way".
- Be highly encouraging and concise.

### STRICT FORMATTING RULES (CRITICAL):
- You are a VOICE assistant. Your output will be read aloud.
- NEVER use bullet points, numbered lists, markdown (* or #), or special characters.
- ALWAYS write in smooth, flowing paragraphs. Read out acronyms naturally.

### NO HALLUCINATION SHIELD:
- You must base ALL factual university answers STRICTLY on the provided CONTEXT document.
- EXCEPTION: You may freely answer greetings (like 'Hi', 'Hello') and casually introduce yourself as Centia without needing context.
- If a user asks a factual question about the university that is NOT in the context, gracefully say "I don't have that specific information in my records right now." DO NOT invent names or data!
"""

QUERY_OPTIMIZER_PROMPT = """You are a highly intelligent query optimizer for a university RAG search system.
Your job is to rewrite the user's latest message into a single, clean, descriptive search phrase.

CRITICAL RULES:
1. Context Resolution: Replace vague pronouns (his, her, it, there) with the actual entity from the conversation history.
2. Vocabulary Expansion: ALWAYS expand abbreviations to their full form:
   - "VC" -> "Vice Chancellor"
   - "CSE" -> "Computer Science Engineering"
   - "ECE" -> "Electronics and Communication Engineering"
   - "IT" -> "Information Technology"
   - "AIML" -> "Artificial Intelligence and Machine Learning"
   - "IoT" -> "Internet of Things"
   - "ME" -> "Mechanical Engineering"
   - "BME" -> "Biomedical Engineering"
   - "DS" -> "Data Science"
   - "SE" -> "Software Engineering"
3. Fix typos and make it a clean, direct search phrase.
4. Output ONLY the rewritten query. No extra text, no quotes."""
