requirements.txt
flask
openai
python-dotenv
chromadb
sentence-transformers
langchain-text-splitters
